% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hfa_xml.R
\name{import.hfa.xml}
\alias{import.hfa.xml}
\title{Import VF from HFA XML}
\usage{
import.hfa.xml(xmlfile, csvfilenamestub, idfunction = function(x) x)
}
\arguments{
\item{xmlfile}{Name of XML file to read}

\item{csvfilenamestub}{Base name of csv files. Protocol name will be appended.}

\item{idfunction}{Function to be applied to patient ID, for instance to correct errors or to de-identify (default: identity function)}
}
\value{
The protocol detected in the input file. (returned invisibly). Side effect of the function is to append the HVF values to the appropriate protocol file (if the file exists, or create a new file).
}
\description{
Import VF from HFA XML
}
\examples{
indir__ <- "S:/HFA Data/HFA2-Peds (Completed)"
outdir__ <- "C:/Users/UM_Ophthy_HSR/Downloads/hfa"
outname__ <- "pupil"

xmlfiles__ <- dir(path=indir__, pattern="\\\\.xml$", full.names = TRUE, ignore.case = TRUE)
nfiles <- length(xmlfiles__)
cat("number of files to process:", nfiles, "\n")
protocols <- character(nfiles)
}
